requirements.txt
Django
django-admin
Pillow
celery
redis
django-celery-results